requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
wheel
matplotlib
numpy==1.20.3
ta-lib
pandas
bravado
websocket-client==0.52.0
hyperopt
pyti
google-api-python-client
oauth2client
python-dotenv